moved get_img_data
